[google_maps_flutter] Widget based polyline support on both platform

+++ packages/google_maps_flutter/lib/src/polyline.dart
@@ -115,8 +115,8 @@ class Polyline {
     List<LatLng> pointsParam,
     Cap startCapParam,
     bool visibleParam,
-    double widthParam,
-    double zIndexParam,
+    int widthParam,
+    int zIndexParam,
     VoidCallback onTapParam,
   }) {
     return Polyline(
